@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core'
-import { Http, Headers, RequestOptions } from '@angular/http'
+import { HttpClient } from '@angular/common/http'
 
 import { Observable } from 'rxjs/Observable'
 import 'rxjs/add/operator/map'
@@ -11,7 +11,7 @@ import { MEAT_API } from './../app.api'
 
 @Injectable()
 export class OrderService {
-  constructor(private cartService: ShoppingCartService, private http: Http) {}
+  constructor(private cartService: ShoppingCartService, private http: HttpClient) {}
 
   itemsValue(): number {
     return this.cartService.total()
@@ -38,15 +38,7 @@ export class OrderService {
   }
 
   checkOrder(order: Order): Observable<string> {
-    const headers = new Headers()
-    headers.append('Content-Type', 'application/json')
-    return this.http
-      .post(
-        `${MEAT_API}/orders`,
-        JSON.stringify(order),
-        new RequestOptions({ headers: headers })
-      )
-      .map(response => response.json())
-      .map(responseOrder => responseOrder.id)
+    return this.http.post<Order>(`${MEAT_API}/orders`, order)
+      .map(order => order.id)
   }
 }
