@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core'
-import { Http } from '@angular/http'
+import { HttpClient, HttpParams } from '@angular/common/http'
 
 import { Observable } from 'rxjs/Observable'
 import 'rxjs/add/operator/map'
@@ -12,33 +12,30 @@ import { MenuItem } from './../restaurant-detail/menu-item/menu-item.model'
 
 @Injectable()
 export class RestaurantsService {
-  constructor(private http: Http) { }
+  constructor(private http: HttpClient) { }
 
   restaurants(search?: string): Observable<Restaurant[]> {
+    let params: HttpParams = undefined
+    if (search) {
+      params = new HttpParams().set('q', search)
+    }
     return this.http
-      .get(`${MEAT_API}/restaurants`, { params: { q: search } })
-      .map(response => response.json())
-      .catch(ErrorHandler.handleError)
+      .get<Restaurant[]>(`${MEAT_API}/restaurants`, { params })
+
   }
 
   restaurantById(id: string): Observable<Restaurant> {
     return this.http
-      .get(`${MEAT_API}/restaurants/${id}`)
-      .map(response => response.json())
-      .catch(ErrorHandler.handleError)
+      .get<Restaurant>(`${MEAT_API}/restaurants/${id}`)
   }
 
   reviewsOfRestaurant(id: string): Observable<any> {
     return this.http
       .get(`${MEAT_API}/restaurants/${id}/reviews`)
-      .map(response => response.json())
-      .catch(ErrorHandler.handleError)
   }
 
   menuOfRestaurant(id: string): Observable<MenuItem[]> {
     return this.http
-      .get(`${MEAT_API}/restaurants/${id}/menu`)
-      .map(response => response.json())
-      .catch(ErrorHandler.handleError)
+      .get<MenuItem[]>(`${MEAT_API}/restaurants/${id}/menu`)
   }
 }
